ValueChangedEventArgsクラスを専用ファイルに分離
